Update lots of stuff (#899)
* Update tons of versions

* Update sbt to 1.3.0

* Fix test compilation

* Travis needs an update too

* Fix another test

* Fix some deprecation warnings

* Sbt 1.3.2

+++ orePlayCommon/app/ore/OreConfig.scala
@@ -200,8 +200,8 @@ final class OreConfig @Inject()(config: Configuration) {
   }
 
   object performance extends ConfigCategory {
-    val raw: Configuration      = root.get[Configuration]("performance")
-    val nioBlockingFibers: Long = raw.get[Long]("nio-blocking-fibers")
+    val raw: Configuration     = root.get[Configuration]("performance")
+    val nioBlockingFibers: Int = raw.get[Int]("nio-blocking-fibers")
   }
 
   app.load()
